This branch was meant to do several things:
- remove unneeded code,
- consolidate some classes together (e.g., game + core + players),
- split main.js and index.js into multiple files, and
- improve synchronization between objects "shared" between server and
  client.

My approach to this last point was to model the objects (game, players,
etc.) as states and change them with updates (diffs). Whenever the
server made an update, it would apply it to its own object and emit the
update for the clients to update theirs. The reason I am dropping this
branch for now is that this approach creates a significantly larger
number of socket emits, reducing performance on the client side
noticeably. I like this idea, but this refactor is turning out to be too
much effort for what it's worth. I'm committing this code in case I want
to come back to it in the future.